+++ ktrain/graph/models.py
@@ -79,7 +79,7 @@ def graph_node_classifier(name, train_data, layer_sizes=[32, 32], verbose=1):
     x_inp, x_out = graphsage_model.build()
     prediction = keras.layers.Dense(units=num_classes, activation=activation)(x_out)
     model = keras.Model(inputs=x_inp, outputs=prediction)
-    model.compile(optimizer="adam", loss=loss_func, metrics=["accuracy"])
+    model.compile(optimizer=U.DEFAULT_OPT, loss=loss_func, metrics=["accuracy"])
     U.vprint("done", verbose=verbose)
     return model